resolve dependency issue
