+++ components/application-mgt/org.wso2.carbon.identity.application.mgt/src/main/java/org/wso2/carbon/identity/application/mgt/ApplicationMgtUtil.java
@@ -18,10 +18,14 @@
 
 package org.wso2.carbon.identity.application.mgt;
 
+import com.fasterxml.jackson.core.JsonProcessingException;
+import com.fasterxml.jackson.databind.ObjectMapper;
 import org.apache.commons.lang.ArrayUtils;
 import org.apache.commons.lang.StringUtils;
 import org.apache.commons.logging.Log;
 import org.apache.commons.logging.LogFactory;
+import org.json.JSONArray;
+import org.json.JSONObject;
 import org.wso2.carbon.CarbonConstants;
 import org.wso2.carbon.base.MultitenantConstants;
 import org.wso2.carbon.base.ServerConfiguration;
@@ -81,6 +85,8 @@ public class ApplicationMgtUtil {
     // Does not allow leading and trailing whitespaces.
     public static final String APP_NAME_VALIDATING_REGEX = "^[a-zA-Z0-9._-]+(?: [a-zA-Z0-9._-]+)*$";
     private static final String SERVICE_PROVIDERS_NAME_REGEX = "ServiceProviders.SPNameRegex";
+    public static final String MASKING_CHARACTER = "*";
+    public static final String MASKING_REGEX = "(?<!^.?).(?!.?$)";
 
     private static Log log = LogFactory.getLog(ApplicationMgtUtil.class);
 
@@ -818,6 +824,101 @@ public class ApplicationMgtUtil {
         }
 
         return ApplicationConstants.DEFAULT_RESULTS_PER_PAGE;
+    }
+
+    /**
+     * Get the application id from service provider object.
+     *
+     * @param serviceProvider Service provider object.
+     * @return Id of the service provider.
+     */
+    public static String getAppId(ServiceProvider serviceProvider) {
+
+        if (serviceProvider != null) {
+            return serviceProvider.getApplicationResourceId();
+        }
+        return StringUtils.EMPTY;
+    }
+
+    /**
+     * Get the application name from service provider object.
+     *
+     * @param serviceProvider Service provider object.
+     * @return Name of the service provider.
+     */
+    public static String getApplicationName(ServiceProvider serviceProvider) {
+
+        if (serviceProvider != null) {
+            return serviceProvider.getApplicationName();
+        }
+        return "Undefined";
+    }
+
+    /**
+     * Get the initiator id.
+     *
+     * @param userName     Username of the initiator.
+     * @param tenantDomain Tenant domain of the initiator.
+     * @return User id of the initiator.
+     */
+    public static String getInitiatorId(String userName, String tenantDomain) {
+
+        String userId = PrivilegedCarbonContext.getThreadLocalCarbonContext().getUserId();
+        if (userId == null) {
+            String userStoreDomain = UserCoreUtil.extractDomainFromName(userName);
+            String username = UserCoreUtil.removeDomainFromName(userName);
+            int tenantId = IdentityTenantUtil.getTenantId(tenantDomain);
+            try {
+                userId = IdentityUtil.resolveUserIdFromUsername(tenantId, userStoreDomain, username);
+            } catch (IdentityException e) {
+               log.error("Error occurred while resolving Id for the user: " + username);
+            }
+        }
+        return userId;
+    }
+
+    /**
+     * Build the service provider JSON string masking the sensitive information.
+     *
+     * @param serviceProvider Service provider object.
+     * @return JSON string of the service provider object.
+     */
+    public static String buildSPData(ServiceProvider serviceProvider) {
+
+        if (serviceProvider == null) {
+            return StringUtils.EMPTY;
+        }
+        try {
+            JSONObject serviceProviderJSONObject =
+                    new JSONObject(new ObjectMapper().writeValueAsString(serviceProvider));
+            JSONObject inboundAuthenticationConfig =
+                    serviceProviderJSONObject.optJSONObject("inboundAuthenticationConfig");
+            if (inboundAuthenticationConfig != null) {
+                JSONArray inboundAuthenticationRequestConfigsArray =
+                        inboundAuthenticationConfig.optJSONArray("inboundAuthenticationRequestConfigs");
+                if (inboundAuthenticationRequestConfigsArray != null) {
+                    for (int i = 0; i < inboundAuthenticationRequestConfigsArray.length(); i++) {
+                        JSONObject requestConfig = inboundAuthenticationRequestConfigsArray.getJSONObject(i);
+                        JSONArray properties = requestConfig.optJSONArray("properties");
+                        if (properties != null) {
+                            for (int j = 0; j < properties.length(); j++) {
+                                JSONObject property = properties.optJSONObject(j);
+                                if (property != null && StringUtils.equalsIgnoreCase("oauthConsumerSecret",
+                                        (String) property.get("name"))) {
+                                    String secret = (String) property.get("value");
+                                    String maskedSecret = secret.replaceAll(MASKING_REGEX, MASKING_CHARACTER);
+                                    property.put("value", maskedSecret);
+                                }
+                            }
+                        }
+                    }
+                }
+            }
+            return serviceProviderJSONObject.toString();
+        } catch (JsonProcessingException e) {
+            log.error("Error while converting service provider object to json.");
+        }
+        return StringUtils.EMPTY;
     }
 
 }
